Add Jest, Playwright and pytest tests (#207)
* Add jest and playwright testing
Fixes #137

* Fix working directory

* Use Jest 26

* Don't lint tests files

* Fix manifest

* Add update snapshots action

* Add documentation

* Add pytest

* Fix cookiecutter

* Print log and stop jupyter lab

* First install jupyter lab

* partial revert to test stuck CI task

* Fix pytest

* Ensure server log redirection

* Move ui tests config in subfolder

* Add documentation for pytest

* Run integration tests in post job using distribution artifact

* Apply review suggestions
- Handle server spin up in Playwright config
- Cache browser download
- Use one artifact
- Don't constrain explicitly Python

* Update check test

* [skip ci] Update documentation

* Remove galata and playwright from main package.json

* Update galata

* Add jest and playwright testing
Fixes #137

* Fix working directory

* Use Jest 26

* Don't lint tests files

* Fix manifest

* Add update snapshots action

* Add documentation

* Add pytest

* Fix cookiecutter

* Print log and stop jupyter lab

* First install jupyter lab

* partial revert to test stuck CI task

* Fix pytest

* Ensure server log redirection

* Move ui tests config in subfolder

* Add documentation for pytest

* Run integration tests in post job using distribution artifact

* Apply review suggestions
- Handle server spin up in Playwright config
- Cache browser download
- Use one artifact
- Don't constrain explicitly Python

* Update check test

* [skip ci] Update documentation

* Remove galata and playwright from main package.json

* Update galata

* Fixed rebase

* Fix prettier

* Fix mac job

* Apply suggestions

* Add configurable for setting up the tests

* Fix syntax

* Fix for new cookiecutter

* Minor fixes

* Conditioned type jest

+++ hooks/post_gen_project.py
@@ -4,11 +4,14 @@ from pathlib import Path
 PROJECT_DIRECTORY = Path.cwd()
 
 
-def remove_path(path: str) -> None:
+def remove_path(path: Path) -> None:
     """Remove the provided path.
     
     If the target path is a directory, remove it recursively.
     """
+    if not path.exists():
+        return
+
     if path.is_file():
         path.unlink()
     elif path.is_dir():
@@ -35,10 +38,21 @@ if __name__ == "__main__":
         for f in (
             "{{ cookiecutter.python_name }}/handlers.py",
             "src/handler.ts",
-            "jupyter-config"
+            "jupyter-config",
+            "conftest.py",
+            "{{ cookiecutter.python_name }}/tests"
         ):
             remove_path(PROJECT_DIRECTORY / f)
 
     if not "{{ cookiecutter.has_binder }}".lower().startswith("y"):
         remove_path(PROJECT_DIRECTORY / "binder")
         remove_path(PROJECT_DIRECTORY / ".github/workflows/binder-on-pr.yml")
+
+    if not "{{ cookiecutter.test }}".lower().startswith("y"):
+        remove_path(PROJECT_DIRECTORY / ".github" / "workflows" / "update-integration-tests.yml")
+        remove_path(PROJECT_DIRECTORY / "src" / "__tests__")
+        remove_path(PROJECT_DIRECTORY / "ui-tests")
+        remove_path(PROJECT_DIRECTORY / "{{ cookiecutter.python_name }}" / "tests")
+        remove_path(PROJECT_DIRECTORY / "babel.config.js")
+        remove_path(PROJECT_DIRECTORY / "conftest.py")
+        remove_path(PROJECT_DIRECTORY / "jest.config.js")
